chore: relocate event-related components to components/events
